+++ src/testing/index.ts
@@ -7,5 +7,3 @@
 
 export * from './interfaces';
 export * from './test';
-export * from './testing-module.builder';
-export * from './testing-module';